pyproject.toml
[project]
name = "rlm-codelens"
version = "0.1.0"
description = "RLM-powered repository analysis using Recursive Language Models with enterprise-grade cost control and security"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Nijesh Kanjinghat", email = "nijesh@example.com"}
]
keywords = ["rlm", "repository-analysis", "github", "openai", "embeddings", "clustering"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "hdbscan>=0.8.33",
    "matplotlib>=3.7.0",
    "networkx>=3.1.0",
    "nltk>=3.8.0",
    "numpy>=1.24.0",
    "openai>=1.0.0",
    "pandas>=2.0.0",
    "plotly>=5.15.0",
    "psycopg2-binary>=2.9.0",
    "pygithub>=2.1.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "rlms>=0.1.0",
    "scikit-learn>=1.3.0",
    "seaborn>=0.12.0",
    "spacy>=3.7.0",
    "sqlalchemy>=2.0.0",
    "tiktoken>=0.12.0",
    "tqdm>=4.65.0",
]

[project.optional-dependencies]
dev = [
    "black>=23.0.0",
    "flake8>=6.0.0",
    "ipykernel>=6.25.0",
    "ipython>=8.0.0",
    "jupyter>=1.0.0",
    "mypy>=1.5.0",
    "pre-commit>=3.4.0",
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.11.0",
    "sphinx>=7.0.0",
    "sphinx-rtd-theme>=1.3.0",
]

[project.scripts]
rlmc = "rlm_codelens.cli:main"

[project.urls]
Homepage = "https://github.com/knijesh/rlm-codelens"
Repository = "https://github.com/knijesh/rlm-codelens"
Issues = "https://github.com/knijesh/rlm-codelens/issues"
Documentation = "https://github.com/knijesh/rlm-codelens#readme"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/rlm_codelens"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.venv
  | build
  | dist
)/
'''

[tool.ruff]
line-length = 88
target-version = "py311"
select = [
    "E",   # pycodestyle errors
    "F",   # Pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "W",   # pycodestyle warnings
]
ignore = [
    "E501",  # line too long (handled by black)
]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
warn_redundant_casts = true
warn_unused_ignores = true
show_error_codes = true
